Implement typed locals
This patch implements support for non-Object typed local bindings,
including fn arguments and type fields, when code is compiled with
`*strict-tags*` bound true or with ^:strict metadata. Strict locals
function as expected in regular, primitive and static fn contexts
as well as in reify and deftype.

As necessary, this patch generates local variable checking and init
code to ensure that tagged locals are correctly checked and stored.

Conditinalizes a large number of interesting checkcasts so that strictly
tagged locals will not emit statically unnecessary type casts.