Download deps in Circle ahead of build time
I'm not 100% sure why, but mvn dependency:go-offline does not download
all of the necessary dependencies to run the build. This means that
Circle's caching isn't very effective, as more deps are downloaded after
the cache is saved.

This patch manually downloads the necessary plugins ahead of time, and
should dramatically speed up build time.

I also needed to update the location for central. See
http://stackoverflow.com/questions/35022961/how-do-i-use-mvn-dependencyget-to-use-https-repo1-maven-org/35023015#35023015
for more details.