Emit better ^:deprecated warnings in refer
This patch tweaks the way that refer emits warnings when working with
deprecated vars and namespaces.

The goal of warnings is to inform users when they're engaging in
behaivors which are unsafe (likely to break in the future) but to reduce
false positives as much as possible so that people will actually read
the warnings rather than turning them off.

- Using a deprecated namespace should emit a warning for the ns, but not
  one for each var. That's overly verbose and likely to swamp a user.

- Using a deprecated var explicitly (eg. via refer or only) is cause for
  a warning because if that var is removed, the refer form will break.

- Referring a deprecated var implicitly (eg. via all) is not cause for a
  warning. Using that var will cause a warning, but generating a warning
  on the referral will emit false positives including on clojure.core
  which are unacceptable. This aproach achieves the goal of not
  generating false positives on clojure.core or other namespaces without
  enshrining some special case code in clojure.core.