Mask out ^:deprecated warnings within dep'd defs
This patch creates a new dynamic context var in the compiler which is
used to track when the form being compiled is marked as deprecated. When
compiling a deprecated form, warnings about the use of deprecated vars
are suppressed.

The justification for this change is that when a var is deprecated, it
should be abandoned/avoided. Consequently it doesn't matter if a
deprecated var makes use of other deprecated functionality so long as
there are no uses of deprecated functionality by non-deprecated
code. Warnings about the use of deprecated vars outside of a deprecated
context are useful, but warnings about the use of deprecated vars within
a deprecated context are just noise. As this is a baked in linter whose
use I want to encourage false positives are more harmful than false
negatives lest users turn off noisey warnings.

The counter argument is that if a deprecated var makes use of other
deprecated vars especially those referred from a library, suppressing
warnings about those uses may lead to breakages in the future if/when
the depended on library vars are removed.