Manually download deps
This patch creates a script, circle-deps.sh, which hashes itself, the
pom and build files, and checks for a flag file in ~/.m2. If the flag is
not found, then a full build is executed in such a way as to be
effectively a dry run, that is to deploy no artifacts and soforth while
using all plugins. This serves to fill the ~/.m2 cache. The flag file is
then created and life continues.

Future builds using the same pom and build files will not have to go
through this double build process. Note that the <version/> tag is
excluded from the hashed text, so that just changing the artifact
version won't impact cache status.