Versioned Namespaces to approximate Modules
This patch creates a concept of versioning in Namespaces. A Namespace's
version simply counts the number of times that the Namespace has been
redefined, that is the ns form has been used.

When a Namespace is redefined, its version number is incremented, and
its imports are reset to the default imports and its mappings are
cleared of anything which is not a default mapping or a mapping of a Var
defined in the reset Namespace. This enables two things.

Firstly, when a Namespace is reloaded its imports/aliases are cleared
allowing users to reuse import qualifications and names when they reload
a Namespace without requiring a restart of the Clojure instance or
generating warnings about name replacement.

Secondly, Vars can make use of this version information to determine
whether they are in sync with the Namespace in which they are defined or
not. This allows vars to give meaningful warnings if they become stale,
and helps users detect and prevent a class of errors where code works at
a repl because it relies on deleted defs which persist, but breaks when
restarted because the deleted defs are not reloaded.

Unfortunately, it is impossible to strategically suppress ns reloading
warnings only during initialization. The concept of initialization is
just too vague, and there are too many strange cases of things being
loaded multiple times to really get a handle on when this warning should
be employed. So no warning is generated when a ns is reset.