Refactor defonce to make use of ^:once
This patch refactors the `defonce` macro so that rather than simply
`def`ing the `Var` in question, instead `Var.intern` is used find or
create the `Var`, and then only if neither `Var.isOnce` nor
`Var.hasRoot` is bound will the `def` of the Var with `^:once` and the
provided body be executed.

This ensures that the defined `Var` will get `^:once` if and only if it
is defined by the `defonce` body, while checking `Var.isOnce` in
addition to `Var.hasRoot`.

This enables the contract that `Var.isOnce` absolutely indicates whether
a `Var` should ever be rebound, enabling Java interop to create `^:once`
`Var`s just as well as Jaunt code can.