Fix overly specific protocols meta test
Previously, this test checked for exact equality of protocol Var
metadata. Since Vars have a metadata values such as ^:once and ^:dynamic
which reflect internal state this test doesn't make a whole lot of sense
since really what it's checking for is that meaningful fields such as
^:name, ^:arglists and ^:doc are set.

Consequently this test adds an `at-least?` function, which checks
essentially that `(= (select-keys r (keys l)) l)`, being that the kvs in
the left map occur exactly in the right map. This allows other metadata
on a Var to exist without breaking the tests.