Set *agent* metadata in core not RT
This patch deals with one instance of the Var metadata state loss
bugs. *agent* is dynamic, has a tag and metadata. Previously, its
metadata was altered in RT with .resetMeta(), which lost the ^:dynamic
flag.

This patch fixes that bug, and moves the metadata and tag initialization
of *agent* into core since it doesn't need to be in RT at all.