Set print-initialized correctly
This patch again fixes metadata updating issues. So that tests can alter
it to access basic printing, print-initialized is supposed to be private
and dynamic. It is initially false valued, but is eventually redefined
to be true once Clojure's printing has been initialized.

This patch fixes the way that print-initialized is redefined so that
pertinent metadata is not lost.