Rework Var to track metadata state correctly
This patch makes Var impement the metadata parts of AReference itself,
so that changes to metadata can be observed. This allows Var to update
the internal status flags corresponding to isPublic, isDynamic, isMacro
etc.

This deals with a class of state errors wherein the metadata of the Var
would not correctly reflect the state of the Var. For instance, since
privacy was decoupled from ^:private, it was possible for a Var to
become `{:private true}` and be .isPrivate() -> False, and vice versa.

Furthermore this patch causes Var to emit warnings when a Var looses its
^:dynamic, ^:once or ^:private bits. This helps capture a class of user
metadata update errors where the Var's metadata is simply reset.

Fixes #42