Suppress stale warnings in redefs, with option
This patch tweaks the compiler to provide support for suppressing stale
var use warnings when a new def of the referenced var is being
compiled.

This masks out a class of false positive warnings emitted when compiling
replacement definitions of stale vars.

So that these warnings can be masked more generally, a compiler option
:warn-on-stale is also provided. When this option is truthy, stale
warnings will be suppressed. As with the other warnings, this option is
implied by :pedantic.