Implement warning for transitive use of stale Vars
This patch modifies the compiler to add code for tracking the set of
Vars used in a given def, and to add that information as metadata to
Vars when they are compiled.

Using this new metadata (`:clojure.core.compiler/uses`), this patch
defines Compiler/getUsedVars (gets the use set from metadata defaulting
to the empty set), Compiler/getReachedVars (computes the reached set by
recursive decent) and Compiler/reachesStaleVars (filter reached for
stale).

When analyzing a Var usage, if that Var reaches any stale Var(s), then a
warning containing the set of reached stale Vars is emitted.