Add use refactor suggestions
This patch causes use to emit a warning containing a rewrite of the use
form to a require form. Does a bunch of legwork to try and rewrite use
forms into a minimal requires. Multiple use forms reaching the same ns
are collapsed. A good effort is also made to format the resulting
warning nicely.

Also adds a warning if renaming is used that renaming is a Bad Idea and
support may be dropped.