Include macros in the use set of resulting fns
This one is kinda iffy, because the macros aren't strictly part of the
use set of the resulting form(s), but rather are required in the
compilation environment of the resulting form. So... they're a
compilation data dependency but not an invocation data dependency.

Maybe this is better split out into its own set? For static linking I
only care about non-macro dependencies, because reloading one of those
will force the reload of all its users. But if a macro used to define a
function changed, then the resulting definition of all macro-derived
code will also change on reload so even though nominal correctness of
existing compiled code may hold you _probably_ want to reload everything
using the macro.