Fix (:refer ...) rewrite generated by refer
Because refer was originally a function, call sites use quoted
arguments. However because in the ns form refer is evaluated at macro
time, (:refer ...) clauses in the ns form need not be quoted.

This patch removes correct but uncalled for unquoting of arguments
passed to the refer* inner function, and explicitly unquotes all
arguments before generating a (:refer ...) rewrite.